Moved CLI arguments into the shared crate.
